chore(rust-sdk): update rustdoc
