+++ src/lib.rs
@@ -5,6 +5,7 @@
 // If the version of rust used is less than v1.63, please uncomment the follow attribute.
 // #![feature(explicit_generic_args_with_impl_trait)]
 #![allow(clippy::vec_init_then_push)]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 //! # Overview
 //!
@@ -79,11 +80,13 @@
 //!
 
 #[cfg(all(feature = "async", target_os = "linux"))]
+#[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
 pub mod r#async;
 #[doc(hidden)]
 pub mod caller;
 #[doc(hidden)]
 #[cfg(feature = "aot")]
+#[cfg_attr(docsrs, doc(cfg(feature = "aot")))]
 mod compiler;
 pub mod config;
 pub mod dock;
@@ -100,7 +103,6 @@ pub mod plugin;
 mod statistics;
 mod store;
 pub mod types;
-#[doc(hidden)]
 pub mod utils;
 #[doc(hidden)]
 pub mod vm;
@@ -110,6 +112,7 @@ pub mod wasi;
 pub use caller::Caller;
 #[doc(inline)]
 #[cfg(feature = "aot")]
+#[cfg_attr(docsrs, doc(cfg(feature = "aot")))]
 pub use compiler::Compiler;
 #[doc(inline)]
 pub use executor::Executor;
@@ -129,18 +132,18 @@ pub use statistics::Statistics;
 #[doc(inline)]
 pub use store::Store;
 #[doc(inline)]
-pub use utils::Driver;
-#[doc(inline)]
 pub use vm::{Vm, VmBuilder};
 
-/// Parses in-memory bytes as either the [WebAssembly Text format](http://webassembly.github.io/spec/core/text/index.html), or a binary WebAssembly module
 pub use wasmedge_types::{
     error, wat2wasm, CompilerOptimizationLevel, CompilerOutputFormat, ExternalInstanceType,
     FuncType, GlobalType, HostRegistration, MemoryType, Mutability, RefType, TableType, ValType,
     WasmEdgeResult,
 };
 
-pub use wasmedge_macro::{async_host_function, host_function};
+#[cfg(all(feature = "async", target_os = "linux"))]
+#[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
+pub use wasmedge_macro::async_host_function;
+pub use wasmedge_macro::host_function;
 
 /// WebAssembly value type.
 pub type WasmValue = wasmedge_sys::types::WasmValue;
@@ -186,28 +189,4 @@ pub trait Engine {
         func_ref: &FuncRef,
         params: impl IntoIterator<Item = WasmValue>,
     ) -> WasmEdgeResult<Vec<WasmValue>>;
-}
-
-/// The version info of WasmEdge core
-pub struct CoreVersion {}
-impl CoreVersion {
-    /// Returns the major version value of WasmEdge core.
-    pub fn major() -> u32 {
-        wasmedge_sys::utils::version_major_value()
-    }
-
-    /// Returns the minor version value of WasmEdge core.
-    pub fn minor() -> u32 {
-        wasmedge_sys::utils::version_minor_value()
-    }
-
-    /// Returns the patch version value of WasmEdge core.
-    pub fn patch() -> u32 {
-        wasmedge_sys::utils::version_patch_value()
-    }
-
-    /// Returns the version string of WasmEdge core.
-    pub fn version_string() -> String {
-        wasmedge_sys::utils::version_string()
-    }
 }
