+++ src/utils.rs
@@ -1,3 +1,4 @@
+//! Defines WasmEdge Driver and CoreVersion types
 use wasmedge_sys::utils;
 
 /// Defines WasmEdge Driver functions
@@ -29,5 +30,29 @@ impl Driver {
         V: AsRef<str>,
     {
         utils::driver_unified_tool(args)
+    }
+}
+
+/// The version info of WasmEdge core
+pub struct CoreVersion {}
+impl CoreVersion {
+    /// Returns the major version value of WasmEdge core.
+    pub fn major() -> u32 {
+        wasmedge_sys::utils::version_major_value()
+    }
+
+    /// Returns the minor version value of WasmEdge core.
+    pub fn minor() -> u32 {
+        wasmedge_sys::utils::version_minor_value()
+    }
+
+    /// Returns the patch version value of WasmEdge core.
+    pub fn patch() -> u32 {
+        wasmedge_sys::utils::version_patch_value()
+    }
+
+    /// Returns the version string of WasmEdge core.
+    pub fn version_string() -> String {
+        wasmedge_sys::utils::version_string()
     }
 }
