+++ src/vm.rs
@@ -161,8 +161,9 @@ impl VmBuilder {
         Ok(vm)
     }
 
-    #[cfg(all(feature = "async", target_os = "linux"))]
     /// Sets the [WasiContext] for the [Vm] to build.
+    #[cfg(all(feature = "async", target_os = "linux"))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
     pub fn with_wasi_context(mut self, wasi_ctx: WasiContext) -> Self {
         self.wasi_ctx = Some(wasi_ctx);
         self
@@ -174,6 +175,7 @@ impl VmBuilder {
     ///
     /// If fail to create, then an error is returned.
     #[cfg(all(feature = "async", target_os = "linux"))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
     pub fn build(mut self) -> WasmEdgeResult<Vm> {
         // executor
         let executor = Executor::new(self.config.as_ref(), self.stat.as_mut())?;
@@ -502,6 +504,7 @@ impl Vm {
     ///
     /// If fail to run the wasm function, then an error is returned.
     #[cfg(all(feature = "async", target_os = "linux"))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
     pub async fn run_func_async(
         &self,
         async_state: &AsyncState,
@@ -580,6 +583,7 @@ impl Vm {
     ///
     /// If fail to run, then an error is returned.
     #[cfg(all(feature = "async", target_os = "linux"))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
     pub async fn run_func_from_module_async<N, A>(
         &mut self,
         async_state: &AsyncState,
@@ -641,6 +645,7 @@ impl Vm {
     ///
     /// If fail to run, then an error is returned.
     #[cfg(all(feature = "async", target_os = "linux"))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
     pub async fn run_func_from_file_async<P, N, A>(
         &mut self,
         async_state: &AsyncState,
@@ -701,6 +706,7 @@ impl Vm {
     ///
     /// If fail to run, then an error is returned.
     #[cfg(all(feature = "async", target_os = "linux"))]
+    #[cfg_attr(docsrs, doc(cfg(all(feature = "async", target_os = "linux"))))]
     pub async fn run_func_from_bytes_async<N, A>(
         &mut self,
         async_state: &AsyncState,
